backend/requirements.txt
djangorestframework-simplejwt
